Add a cast to eliminate a compiler warning.
std::vector returns a size_t from size(). We assign it to a UINT which results in a warning. Adding a cast to make that warning go away.

+++ source/UIData/CodeGen/CxInstantiatorGenerator.cs
@@ -869,7 +869,7 @@ public:
     {
         if (count != nullptr)
         {
-            *count = m_sources.size();
+            *count = (UINT)m_sources.size();
         }
 
         return S_OK;
